changing phonenumber from int to bigint

+++ db/migrate/20131130200613_add_phone_to_users.rb
@@ -1,6 +1,6 @@
 class AddPhoneToUsers < ActiveRecord::Migration
   def change
-  	    add_column :users, :phoneNumber, :integer
+  	    add_column :users, :phoneNumber, :bigint
   	    add_column :users, :provider, :string
   end
 end
